requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
astor==0.8.1
attrs==25.4.0
beautifulsoup4==4.14.2
blake3==1.0.8
boto3==1.40.67
botocore==1.40.67
brotli==1.2.0
cachetools==6.2.1
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
coloredlogs==15.0.1
colorlog==6.10.1
compressed-tensors==0.11.0
contourpy==1.3.3
cryptography==46.0.3
cupy-cuda12x==13.6.0
cycler==0.12.1
datasets==4.4.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
doclayout_yolo==0.0.4
einops==0.8.1
email-validator==2.3.0
fast-langdetect==0.2.5
fastapi==0.121.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastrlock==0.8.3
fasttext-predict==0.9.2.4
ffmpy==0.6.4
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gguf==0.17.1
gradio==5.49.1
gradio_client==1.13.3
gradio_pdf==0.0.22
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-retries==0.4.5
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
interegular==0.3.3
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
json_repair==0.52.4
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
magika==0.6.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mineru_vl_utils==0.1.15
mistral_common==1.8.5
mlx==0.29.3
mlx-lm==0.28.3
mlx-vlm==0.3.5
modelscope==1.31.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.5
ninja==1.13.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.7.1
openai-harmony==0.0.8
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.4
outlines_core==0.2.11
packaging==25.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post6
pdfminer.six==20250506
pdftext==0.6.3
pillow==11.3.0
polars==1.35.1
polars-runtime-32==1.35.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pybase64==1.4.2
pyclipper==1.3.0.post6
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.10
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.51.1
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
robust-downloader==0.0.2
rpds-py==0.28.0
ruff==0.14.3
s3transfer==0.14.0
safehttpx==0.1.7
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.43.0
setproctitle==1.3.7
setuptools==79.0.1
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
starlette==0.49.3
stringzilla==4.2.3
sympy==1.14.0
thop==0.1.1.post2209072238
tifffile==2025.10.16
tiktoken==0.12.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.1
triton==3.4.0
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.225
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
